+++ packages/terra-core-docs/src/terra-dev-site/test/visually-hidden-text/DefaultVisuallyHiddenText.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import VisuallyHiddenText from 'terra-visually-hidden-text';
+
+export default () => (
+  <p>
+    No Screen Reader text is added on this page, and should be blank other than this line.
+    <VisuallyHiddenText />
+  </p>
+);