+++ packages/terra-text/src/terra-dev-site/test/text/DefaultText.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Text from '../../../Text';
-
-export default () => <Text id="text">Default</Text>;